requirements.txt
altgraph==0.17.4
asgiref==3.6.0
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
dj-stripe==2.8.1
Django==4.2.1
django-allauth==0.58.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
Eel==0.16.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mutagen==1.46.0
oauthlib==3.2.2
packaging==23.2
pefile==2023.2.7
Pillow==10.0.0
pycparser==2.21
pycryptodomex==3.17
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyJWT==1.7.1
pyparsing==3.1.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
self==2020.12.3
sqlparse==0.4.4
stripe==4.2.0
tzdata==2023.3
urllib3==2.0.4
websockets==11.0.3
whichcraft==0.6.1
yt-dlp==2023.3.4
zope.event==5.0
zope.interface==6.1
